Fixes issue #33
