+++ emulinker/conf/emulinker.cfg
@@ -65,7 +65,7 @@ game.defaultAutoFireSensitivity=0
  # MASTER LIST CONFIGURATION
  # =========================
  # Server name to display in the master lists 
-masterList.serverName=New EmulinkerSF-netsma Server
+masterList.serverName=Local Dev Server
  # Location of the server to display in the master lists
 masterList.serverLocation=Unknown
  # Website of the server to display in the master lists
