+++ emulinker/src/main/java/org/emulinker/kaillera/model/impl/KailleraServerImpl.kt
@@ -77,9 +77,6 @@ class KailleraServerImpl
     return gamesMap[gameID]
   }
 
-  override val numUsers = usersMap.size
-  override val numGames = gamesMap.size
-
   fun getNumGamesPlaying(): Int {
     var count = 0
     for (game in games) {
@@ -105,7 +102,7 @@ class KailleraServerImpl
 
   override fun toString(): String {
     return String.format(
-        "KailleraServerImpl[numUsers=%d numGames=%d isRunning=%b]", numUsers, numGames, running)
+        "KailleraServerImpl[numUsers=%d numGames=%d isRunning=%b]", users.size, games.size, running)
   }
 
   @Synchronized
@@ -155,7 +152,7 @@ class KailleraServerImpl
   @Throws(ServerFullException::class, NewConnectionException::class)
   override fun newConnection(
       clientSocketAddress: InetSocketAddress?, protocol: String?, listener: KailleraEventListener?
-  ): KailleraUser? {
+  ): KailleraUser {
     // we'll assume at this point that ConnectController has already asked AccessManager if this IP
     // is banned, so no need to do it again here
     logger
@@ -258,7 +255,7 @@ class KailleraServerImpl
               "KailleraServerImpl.LoginDeniedPingTooHigh", user.ping.toString() + " > " + maxPing))
     }
     if (access == AccessManager.ACCESS_NORMAL &&
-        allowedConnectionTypes[user.connectionType.toInt()] == false) {
+        !allowedConnectionTypes[user.connectionType.toInt()]) {
       logger
           .atInfo()
           .log(
@@ -310,9 +307,7 @@ class KailleraServerImpl
     if (access == AccessManager.ACCESS_NORMAL &&
         flags.maxClientNameLength > 0 &&
         user.clientType!!.length > maxClientNameLength) {
-      logger
-          .atInfo()
-          .log(user.toString() + " login denied: Client Name Length > " + maxClientNameLength)
+      logger.atInfo().log("$user login denied: Client Name Length > $maxClientNameLength")
       usersMap.remove(userListKey)
       throw UserNameException(getString("KailleraServerImpl.LoginDeniedEmulatorNameTooLong"))
     }
@@ -612,7 +607,7 @@ class KailleraServerImpl
         logger.atWarning().log("$user create game denied: Flood: $romName")
         throw FloodException(getString("KailleraServerImpl.CreateGameDeniedFloodControl"))
       }
-      if (flags.maxGames > 0 && numGames >= flags.maxGames) {
+      if (flags.maxGames > 0 && games.size >= flags.maxGames) {
         logger
             .atWarning()
             .log(
